requirements.txt
grpcio==1.62.2
grpcio-tools==1.62.2
protobuf==4.25.3
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-grpc==0.42b0
prometheus_client==0.25.0
